Extract the method for getting user info
