+++ src/flickypedia/types/wikimedia.py
@@ -1,11 +1,6 @@
 import typing
 
 
-class UserInfo(typing.TypedDict):
-    id: str
-    name: str
-
-
 class ShortCaption(typing.TypedDict):
     language: str
     text: str
