+++ tests/apis/wikimedia/test_user_methods.py
@@ -1,0 +1,7 @@
+from flickypedia.apis import WikimediaApi
+
+
+def test_get_userinfo(wikimedia_api: WikimediaApi) -> None:
+    info = wikimedia_api.get_userinfo()
+
+    assert info == {"id": 829939, "name": "Alexwlchan"}
